- added utils.py - renamed model.py as arraymodel.py - splited view.py in arraywidget.py, editor.py and comparator.py
